+++ core/docs/changelog.md
@@ -2,6 +2,9 @@
 ### Fixes
 * Correct links now that `kurtosis-libs` is renamed to `kurtosis-testsuite-starter-pack`
 
+### Features
+* The API container will now shut down all the containers in its network as it shuts down, which is a step towards enclaves being independent of the testing framework
+
 # 1.18.0
 ### Fixes
 * Updated copyright notice to 2021, with entity as Kurtosis Technologies Inc.